move crdt sync service into LcmCrdt
