+++ backend/FwLite/LcmCrdt/LcmCrdt.csproj
@@ -18,6 +18,8 @@
         <PackageReference Include="linq2db.SQLite" Version="5.4.1" />
         <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="8.0.0"/>
         <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="8.0.0" />
+        <PackageReference Include="Refit" Version="7.1.2"/>
+        <PackageReference Include="Refit.HttpClientFactory" Version="7.1.2"/>
     </ItemGroup>
 
     <ItemGroup>
