+++ backend/FwLite/LcmCrdt/LcmCrdtKernel.cs
@@ -6,6 +6,7 @@ using SIL.Harmony.Changes;
 using LcmCrdt.Changes;
 using LcmCrdt.Changes.Entries;
 using LcmCrdt.Objects;
+using LcmCrdt.RemoteSync;
 using LinqToDB;
 using LinqToDB.AspNet.Logging;
 using LinqToDB.Data;
@@ -14,6 +15,8 @@ using LinqToDB.Mapping;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using Refit;
 
 namespace LcmCrdt;
 
@@ -33,6 +36,17 @@ public static class LcmCrdtKernel
         services.AddScoped<CurrentProjectService>();
         services.AddSingleton<ProjectContext>();
         services.AddSingleton<ProjectsService>();
+
+        services.AddHttpClient();
+        services.AddSingleton<RefitSettings>(provider => new RefitSettings
+        {
+            ContentSerializer = new SystemTextJsonContentSerializer(new(JsonSerializerDefaults.Web)
+            {
+                TypeInfoResolver = provider.GetRequiredService<IOptions<CrdtConfig>>().Value
+                    .MakeJsonTypeResolver()
+            })
+        });
+        services.AddSingleton<CrdtHttpSyncService>();
         return services;
     }
