+++ backend/FwLite/LocalWebApp/LocalAppKernel.cs
@@ -40,16 +40,6 @@ public static class LocalAppKernel
             {
                 jsonOptions.PayloadSerializerOptions.TypeInfoResolver = crdtConfig.Value.MakeJsonTypeResolver();
             });
-        services.AddHttpClient();
-        services.AddSingleton<RefitSettings>(provider => new RefitSettings
-        {
-            ContentSerializer = new SystemTextJsonContentSerializer(new(JsonSerializerDefaults.Web)
-            {
-                TypeInfoResolver = provider.GetRequiredService<IOptions<CrdtConfig>>().Value
-                    .MakeJsonTypeResolver()
-            })
-        });
-        services.AddSingleton<CrdtHttpSyncService>();
         return services;
     }
