+++ backend/FwLite/LocalWebApp/LocalWebApp.csproj
@@ -24,8 +24,6 @@
         <PackageReference Include="Microsoft.AspNetCore.SignalR.Client" Version="8.0.10" />
         <PackageReference Include="Microsoft.Extensions.FileProviders.Embedded" Version="8.0.10" />
         <PackageReference Include="Microsoft.Identity.Client.Extensions.Msal" Version="4.64.0" />
-        <PackageReference Include="Refit" Version="7.1.2" />
-        <PackageReference Include="Refit.HttpClientFactory" Version="7.1.2" />
         <PackageReference Include="Swashbuckle.AspNetCore" Version="6.7.3" />
         <PackageReference Include="System.Reactive" Version="6.0.1" />
     </ItemGroup>
